+++ src/main/java/org/amoseman/spaceengineershelper/model/AggregatedCost.java
@@ -7,16 +7,16 @@ import java.util.Set;
 
 public class AggregatedCost {
     private final Resource resource;
-    private final HashMap<String, Long> cost;
+    private final HashMap<String, Double> cost;
 
     public AggregatedCost(Resource resource) {
         this.resource = resource;
         this.cost = new HashMap<>();
     }
 
-    public void add(long multiplier, Resource resource) {
+    public void add(double multiplier, Resource resource) {
         String key = resource.getName();
-        long amount = resource.getAmount() * multiplier;
+        double amount = resource.getAmount() * multiplier;
         if (cost.containsKey(key)) {
             amount += cost.get(key);
         }
